requirements.txt
python-docx==0.8.11
PyPDF2==2.12.1
pypandoc==1.15
PyMuPDF>=1.25.0
Pillow>=11.0.0
flask==2.2.3
Flask-Reuploaded==1.3.0
gitpython==3.1.37
python-dotenv==1.0.0
requests==2.28.2
tqdm==4.65.0
Werkzeug==2.2.3
markdown==3.4.3
beautifulsoup4==4.12.0 